+++ src/reveal/reproduction/__init__.py
@@ -1,0 +1,7 @@
+"""Vulnerability reproduction support."""
+
+from reveal.reproduction.base import PocGenerator
+
+__all__ = [
+    "PocGenerator",
+]